Move metadata_mapping.py to handles.py.
